@@ -12,33 +12,33 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "github:angular/animations-builds",
-    "@angular/cdk": "github:angular/cdk-builds",
-    "@angular/common": "~7.0.0",
-    "@angular/compiler": "~7.0.0",
-    "@angular/core": "~7.0.0",
-    "@angular/forms": "~7.0.0",
+    "@angular/animations": "^8.0.0",
+    "@angular/cdk": "8.0.1",
+    "@angular/common": "~8.0.0",
+    "@angular/compiler": "~8.0.0",
+    "@angular/core": "~8.0.0",
+    "@angular/forms": "~8.0.0",
     "@angular/http": "~7.0.0",
-    "@angular/material": "github:angular/material2-builds",
-    "@angular/platform-browser": "~7.0.0",
-    "@angular/platform-browser-dynamic": "~7.0.0",
-    "@angular/router": "~7.0.0",
+    "@angular/material": "8.0.1",
+    "@angular/platform-browser": "~8.0.0",
+    "@angular/platform-browser-dynamic": "~8.0.0",
+    "@angular/router": "~8.0.0",
     "core-js": "^2.5.4",
     "express": "^4.16.4",
     "ngx-bootstrap": "^3.1.1",
     "ngx-markdown": "^7.0.0",
     "ngx-markdown-editor": "^1.2.0",
-    "rxjs": "~6.3.3",
-    "zone.js": "~0.8.26"
+    "rxjs": "~6.5.2",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.10.0",
-    "@angular/cli": "~7.0.4",
-    "@angular/compiler-cli": "~7.0.0",
-    "@angular/language-service": "~7.0.0",
-    "@types/node": "~8.9.4",
+    "@angular-devkit/build-angular": "^0.800.2",
+    "@angular/cli": "~8.0.2",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/language-service": "~8.0.0",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
     "codelyzer": "~4.5.0",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
@@ -50,7 +50,7 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.1.1"
+    "typescript": "~3.4.5"
   },
   "engines": {
     "node": "8.11.4",
